220509 aftrn Ex2420 change

+++ main/src/solvedAc/Bronze4/Ex2420.java
@@ -6,11 +6,11 @@ public class Ex2420 {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
         System.out.print("첫 서브도메인: ");
-        int a = scanner.nextInt();
+        Long a = scanner.nextLong();
         System.out.print("두번째 서브도메인: ");
-        int b = scanner.nextInt();
+        Long b = scanner.nextLong();
 
-        int abNum = 0;
+        Long abNum = 0L;
         abNum = Math.abs(a - b);
 
         System.out.println(abNum);
